+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased] - (Someday)
 
 * Lowered the water fill cycle limit
+* Allow water flow across flat tiles
 
 ## [0.2.3] - (2024-08-31)