+++ changelog.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - (Someday)
 
+## [0.2.4] - (2024-09-01)
+
 * Lowered the water fill cycle limit
 * Allow water flow across flat tiles
 * Allow override of glacier and wetland biome criteria
@@ -35,7 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial public release.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/nms-scribe/cosmopoeia/v0.2.3...HEAD
+[Unreleased]: https://github.com/nms-scribe/cosmopoeia/v0.2.4...HEAD
+[0.2.4]: https://github.com/nms-scribe/cosmopoeia/v0.2.3...v0.2.4
 [0.2.3]: https://github.com/nms-scribe/cosmopoeia/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/assert-rs/predicates-rs/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/assert-rs/predicates-rs/compare/v0.2.0...v0.2.1